+++ hjn20160520/Common/Printer.cs
@@ -100,7 +100,7 @@ namespace hjn20160520.Common
 
             sb.Append("  商品编号" + "\t" + "品名" + "\t" + "数量" + "\t" + "金额" + "\n");
 
-            int count_temp = 0; //合计数量
+            decimal count_temp = 0.00m; //合计数量
             //品名设置每7字换一行
             foreach (var item in goodsList)
             {
